+++ libs/react-components/src/lib/components/button/button.spec.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import { ThemeProvider } from 'styled-components';
 import { lightTheme } from '../../themes';
-import Button, { ButtonProps } from './button';
+import { Button, ButtonProps } from './button';
 import '@testing-library/jest-dom/extend-expect';
 import renderer from 'react-test-renderer';
 import 'jest-styled-components';