+++ libs/react-components/src/lib/components/button/button.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import Button from './button';
+import { Button } from './button';
 
 export default {
   title: 'Theme/Button',