+++ libs/react-components/src/lib/components/index.ts
@@ -1,1 +1,5 @@
-export * from './button/button';
+export * from './forms';
+export * from './button';
+export * from './link';
+export * from './section';
+export * from './typography';