+++ libs/react-components/src/lib/components/link/changelog.md
@@ -2,4 +2,4 @@
 
 ## 0.1.0
 
-- [MINOR] Initial component creation. Support for
+- [MINOR] Initial component creation. Support for secondary and inverse variants, turning off `text-decoration` and for using with HTML button elements